Set of sensor lists for a beaked_whale

Data is from a _Mesoplodon densirostris_ with tag ID md13_134a. The device used was a DTAG3 and it was deployed at 2013-05-14 12:42:00 in El Hierro, Canary Islands, Spain.

Usage

beaked_whale

Format

A set of sensor lists:

A

sensor list containing a triaxial acceleration matrix sampled at 25 Hz

M

sensor list containing a triaxial magnetometer matrix sampled at 25 Hz

P

sensor list containing a pressure (depth) vector sampled at 25 Hz