Extract a sub-sample of data

This function is used to extract a sub-sample of data from a vector or matrix.

Usage

extract(x, sampling_rate, tst, ted)

Arguments

x

A vector or matrix of measurements. If x is a matrix, each column is treated as a separate measurement vector.

sampling_rate

the sampling rate in Hz of the data in x.

tst

Defines the start time in seconds of the interval to be extracted from x.

ted

Defines the end time in seconds of the interval to be extracted from x.

Value

X: A matrix containing a sub-sample of x. X has the same number of columns as x. The length of the sub-sample will be round(sampling_rate*(tend-tstart)) samples.

Note

Output sampling rate is the same as the input sampling rate.

If either tst or ted are beyond the length of x, non-existing samples will be replaced with NaN in X.

Examples

BW <- beaked_whale
extract(x = BW$A$data, sampling_rate = BW$A$sampling_rate, tst = 3, ted = 100)
#>               [,1]          [,2]      [,3]
#>  [1,]  0.359353567  0.0740137342 0.8712770
#>  [2,] -0.077931739 -0.0819311791 0.9551228
#>  [3,] -0.301606964 -0.2120884869 0.9020090
#>  [4,] -0.246412954 -0.1765318686 0.9647387
#>  [5,] -0.159174577 -0.0977487930 1.0446557
#>  [6,] -0.147319041 -0.0669413730 0.9646837
#>  [7,] -0.098077929 -0.0245823711 0.9554624
#>  [8,]  0.009598672 -0.0378381947 1.0513772
#>  [9,]  0.219067566  0.0208593498 1.0337585
#> [10,]  0.407521613  0.0837601168 0.8521666
#> [11,]  0.237916597 -0.0006010238 0.8910755
#> [12,] -0.160812378  0.0082994583 1.0128137
#> [13,] -0.326421787 -0.0233040953 0.9297769
#> [14,] -0.291309381 -0.0924453263 0.9603509
#> [15,] -0.350049094 -0.0611884309 0.9637004
#> [16,] -0.283548018 -0.0837481949 0.9436157
#> [17,] -0.129484069 -0.0447882291 0.9952420
#> [18,] -0.086746174 -0.0694651113 0.9942092
#> [19,] -0.042561297 -0.0766761343 1.0015410
#> [20,]  0.133620569 -0.0182216757 1.0344031
#> [21,]  0.378364184 -0.0329534169 0.9503804
#> [22,]  0.311084178  0.0114752462 0.8499618
#> [23,] -0.061575390  0.0263671180 0.9448094
#> [24,] -0.275594818  0.0100324648 1.0059493
#> [25,] -0.255330560 -0.0074074994 0.9867141
#> [26,] -0.280400819 -0.0122309443 0.9534131
#> [27,] -0.243873617 -0.0323334034 0.9539954
#> [28,] -0.033006679 -0.0469847267 1.0195104
#> [29,]  0.136211494  0.0234239363 1.0312311
#> [30,]  0.387141901  0.0330023097 0.9417306
#> [31,]  0.470581501 -0.0007027594 0.8231513
#> [32,]  0.103192038  0.0130153486 0.9299108
#> [33,] -0.145103562 -0.0069316867 1.0513772
#> [34,] -0.225676198 -0.0826042329 0.9359654
#> [35,] -0.316131516 -0.1072972149 0.9119763
#> [36,] -0.212802737 -0.0832102448 1.0057865
#> [37,] -0.094516307 -0.0589526910 1.0274974
#> [38,]  0.013638832 -0.0268900412 1.0068812
#> [39,]  0.288855733  0.0369511439 0.9694065
#> [40,]  0.457368690  0.0497775504 0.8747854
#> [41,]  0.194268563 -0.0310481060 0.8877674
#> [42,] -0.129789890 -0.1102572056 1.0390674
#> [43,] -0.229179017 -0.0344801723 0.9977547
#> [44,] -0.325044625 -0.0377363209 0.8774684
#> [45,] -0.240522696 -0.1372373537 0.9608579
#> [46,] -0.019108451 -0.0137549226 1.0331258
#> [47,]  0.116596164  0.0012622162 1.0415130
#> [48,]  0.342587179 -0.0639029574 0.9666331
#> [49,]  0.409972471  0.0407577288 0.8486385
#> [50,]  0.076006127  0.0259095513 0.9121590
#> [51,] -0.182267232 -0.1437875212 0.9924554
#> [52,] -0.198164234 -0.1491726719 1.0004716
#> [53,] -0.269926841 -0.0120613156 0.9667695
#> [54,] -0.273809451  0.0040281400 0.9475897
#> [55,] -0.092236949 -0.0232454676 0.9898177
#> [56,]  0.100726385 -0.0383335459 1.0290936
#> [57,]  0.325774301 -0.0491489633 0.9833272
#> [58,]  0.418483685  0.0193166202 0.8876654
#> [59,]  0.131091879  0.0479487757 0.9206949
#> [60,] -0.226798313  0.0197961137 0.9578922
#> [61,] -0.331655473 -0.0036184684 0.9396418
#> [62,] -0.274005911 -0.0168312160 0.9804647
#> [63,] -0.290539927 -0.0172855342 0.9845927
#> [64,] -0.305151044 -0.0045026034 0.9636764
#> [65,] -0.177366766 -0.0105339894 0.9645688
#> [66,] -0.074895988 -0.0555228038 0.9863405
#> [67,]  0.078467083 -0.0381752894 1.0288020
#> [68,]  0.342561649 -0.0276366769 1.0111551
#> [69,]  0.491536121 -0.0100946591 0.8667591
#> [70,]  0.265047636  0.0225219441 0.8238848
#> [71,] -0.114906269  0.0160899734 0.9948660
#> [72,] -0.307208667 -0.0182221398 0.9906380
#> [73,] -0.342640626 -0.0923519953 0.8905236
#> [74,] -0.221588763 -0.0919093664 0.9979014
#> [75,] -0.031502986 -0.0543362685 1.0519329
#> [76,]  0.058369870  0.0031200269 0.9791157
#> [77,]  0.114635939  0.0588836512 0.9891672
#> [78,]  0.375773230 -0.0298213661 0.9496757
#> [79,]  0.443885045 -0.0718207878 0.8737444
#> [80,]  0.018616023  0.0024824369 0.9453466
#> [81,] -0.245656922 -0.0496538435 0.9682467
#> [82,] -0.230977573 -0.0099943369 0.9605742
#> [83,] -0.332988144  0.0406231944 0.9507593
#> [84,] -0.408680700 -0.0576346165 0.9135399
#> [85,] -0.257934630  0.0275143864 0.9833678
#> [86,] -0.270586140  0.0554195365 0.9765012
#> [87,] -0.416996833 -0.0082451865 0.8864101
#> [88,] -0.290355870  0.0576799550 0.9406754
#> [89,] -0.152320531  0.1046671518 1.0099117
#> [90,] -0.152273619  0.0378132093 1.0078762
#> [91,] -0.121874274 -0.0389983913 0.9891947
#> [92,] -0.007790740 -0.0078124526 0.9897129
#> [93,]  0.161962424 -0.0215814825 0.9937589
#> [94,]  0.299806369 -0.0880273105 0.9874128
#> [95,]  0.435064510 -0.0189969897 0.9199355
#> [96,]  0.437887685  0.0652293476 0.8063729
#> [97,]  0.084631782  0.0431209470 0.9311981