Plot coloured line(s) in 3 dimensions with plot_ly

This function is used to plot three dimensional lines with segments colored. It may be just as simple to use plotly::plot_ly() directly.

Usage

col_line3(x, y, z = 0, c, ...)

Arguments

x

name of object or variable containing data for x axis

y

name of object or variable containing data for y axis

z

name of object or variable containing data for z axis

c

name of object or variable by which to color

...

Additional inputs for plot_ly()

Value

a plot_ly() graphics object

Note

x, y, z and c must all be the same size vectors. The color axis will by default span the range of values in c, i.e., caxis will be c(min(min(c)), max(max(c))).

See also

Examples

col_line3(1:20, 1:20, 1:20, 1:20)